feat: Items.jsx React Query로 전환 및 Items/index.jsx 삭제
